@@ -7,15 +7,21 @@ class App extends React.Component {
 
   constructor() {
     super();
-    
+
     this.state = {
       latestTweets: []
     };
   }
 
-  // TODO: componentWillMount()
-  // TODO: componentDidMount()
-  // TODO: componentWillUnmount()
+  componentWillMount(){
+    this.fetchTweets()
+  }
+
+  componentDidMount(){
+    this.startInterval();
+  }
+
+
 
   startInterval = () => {
     this.interval = setInterval(this.fetchTweets, 2000);
@@ -28,6 +34,10 @@ class App extends React.Component {
     this.setState({
       latestTweets: newTweets
     });
+  }
+
+  componentWillUnmount(){
+    this.cleanUpInterval()
   }
 
   render() {
